Code-share clock tower EntityClouds between NZ1 and RNZ1
