+++ src/st/rnz1/e_clouds.c
@@ -1,4 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "rnz1.h"
 
-INCLUDE_ASM("st/rnz1/nonmatchings/e_clouds", EntityClouds);
+// the reverse clock tower is upside down: the cloud band grows downward
+#define CLOUDS_INVERTED
+
+#include "../e_clock_tower_clouds.h"
